refactor: 話者状態 APIs を Router へ移動
