+++ voicevox_engine/app/routers/speaker.py
@@ -3,12 +3,13 @@
 import base64
 import json
 from pathlib import Path
-from typing import Callable, Literal
+from typing import Annotated, Callable, Literal
 
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, Query, Response
 from pydantic import parse_obj_as
 
 from voicevox_engine.core.core_adapter import CoreAdapter
+from voicevox_engine.metas.Metas import StyleId
 from voicevox_engine.metas.MetasStore import MetasStore, filter_speakers_and_styles
 from voicevox_engine.model import Speaker, SpeakerInfo
 
@@ -163,5 +164,35 @@ def generate_router(
             speaker_or_singer="singer",
             core_version=core_version,
         )
+
+    @router.post("/initialize_speaker", status_code=204, tags=["その他"])
+    def initialize_speaker(
+        style_id: Annotated[StyleId, Query(alias="speaker")],
+        skip_reinit: Annotated[
+            bool,
+            Query(
+                description="既に初期化済みのスタイルの再初期化をスキップするかどうか",
+            ),
+        ] = False,
+        core_version: str | None = None,
+    ) -> Response:
+        """
+        指定されたスタイルを初期化します。
+        実行しなくても他のAPIは使用できますが、初回実行時に時間がかかることがあります。
+        """
+        core = get_core(core_version)
+        core.initialize_style_id_synthesis(style_id, skip_reinit=skip_reinit)
+        return Response(status_code=204)
+
+    @router.get("/is_initialized_speaker", response_model=bool, tags=["その他"])
+    def is_initialized_speaker(
+        style_id: Annotated[StyleId, Query(alias="speaker")],
+        core_version: str | None = None,
+    ) -> bool:
+        """
+        指定されたスタイルが初期化されているかどうかを返します。
+        """
+        core = get_core(core_version)
+        return core.is_initialized_style_id_synthesis(style_id)
 
     return router
